refactor: 잘못된 인자 전역 예외 처리 변경
- InvalidArgumentException 커스텀 예외 클래스 추가
