+++ src/main/java/roomescape/global/exception/InvalidArgumentException.java
@@ -1,0 +1,7 @@
+package roomescape.global.exception;
+
+public class InvalidArgumentException extends RuntimeException {
+    public InvalidArgumentException(String message) {
+        super(message);
+    }
+}
